Version 0.21 (VariableTable removed)
+ Replaced custom control variable table with DataGridView due to error
when loading a lot of variable data.
+ Added scroll to current control in element viewer when a type change
occurs.
+ Updated and expanded UML diagrams.
+ Prepared screenshots for thesis theory.
+ Fixed a minor bug with Simulation Start.

+++ bark_GUI/XmlHandling/XML_Handler.cs
@@ -117,16 +117,20 @@ namespace bark_GUI.XmlHandling
             // Save on temporary file and validate it before saving.
             var tmpFile = Settings.Default.PathSamples + "\\tmp.brk";
 
-            var tmpLastSavedXml = _lastSavedXml;
-            Save(tmpFile);
-            _lastSavedXml = tmpLastSavedXml;
-
-            if (!_ValidateXml(tmpFile))
+            // Avoid infinite loop
+            if (tmpFile != filepath)
             {
+                var tmpLastSavedXml = _lastSavedXml;
+                Save(tmpFile);
+                _lastSavedXml = tmpLastSavedXml;
+
+                if (!_ValidateXml(tmpFile))
+                {
+                    File.Delete(tmpFile);
+                    return false;
+                }
                 File.Delete(tmpFile);
-                return false;
             }
-            File.Delete(tmpFile);
 
 
             // Check for empty path.
